Setup Backend Specs
Previously, we had the application running only on one environment - the development env. These changes allow for creation of an instance of the Express app with a different DB for any environment.

This is just a pre-requisite for writing Node.js tests since we'll need to spin up a separate server for tests.

Author:
@Harrisonkamau